refactor(HistoryListPanel): update import paths for HistoryListPanelCol to use centralized types
