+++ src/components/panels/History/HistoryListPanelExportCsv.vue
@@ -15,7 +15,7 @@ import BaseMixin from '@/components/mixins/base'
 import { ServerHistoryStateJob } from '@/store/server/history/types'
 import { formatFilesize } from '@/plugins/helpers'
 import { mdiDatabaseExportOutline } from '@mdi/js'
-import { HistoryListPanelCol } from '@/components/panels/HistoryListPanel.vue'
+import { HistoryListPanelCol } from '@/store/server/history/types'
 import HistoryMixin from '@/components/mixins/history'
 
 @Component
