+++ src/store/server/history/types.ts
@@ -90,6 +90,16 @@ export interface ServerHistoryStateAllPrintStatusEntry {
 
 export type HistoryStatsValueNames = 'jobs' | 'filament' | 'time'
 
+export interface HistoryListPanelCol {
+    text: string
+    value: string
+    align: string
+    configable: boolean
+    visible: boolean
+    filterable?: boolean
+    outputType?: string
+}
+
 export interface ServerHistoryStateJobWithCount extends ServerHistoryStateJob {
     count: number
 }
